Consolidate `AbstractMysqlAdapter` and `Mysql2Adapter`
